+++ lib/vega/lite_chart.rb
@@ -2,10 +2,10 @@ module Vega
   class LiteChart < BaseChart
     # https://vega.github.io/vega-lite/docs/spec.html
     scalar_methods \
-      :background, :padding, :autosize, :title, :name, :description, :width, :height, :mark
+      :background, :padding, :autosize, :title, :name, :description, :width, :height, :mark, :spec
 
     hash_methods \
-      :config, :usermeta, :projection, :datasets, :encoding
+      :config, :usermeta, :projection, :datasets, :encoding, :repeat, :resolve, :selection
 
     array_methods \
       :transform, :layer