+++ lib/vega/lite_chart.rb
@@ -2,7 +2,7 @@ module Vega
   class LiteChart < BaseChart
     # https://vega.github.io/vega-lite/docs/spec.html
     scalar_methods \
-      :background, :padding, :autosize, :title, :name, :description, :width, :height, :mark
+      :background, :padding, :autosize, :title, :name, :description, :width, :height, :mark, :spec
 
     hash_methods \
       :config, :usermeta, :projection, :datasets, :encoding, :repeat, :resolve, :selection
@@ -20,5 +20,14 @@ module Vega
       self
     end
     immutable_method :data
+
+    undef spec
+    def spec(*args)
+      if args.empty?
+        @spec
+      else
+        dup.spec!(*args)
+      end
+    end
   end
 end