+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 (unreleased)
 
+- Added `repeat`, `resolve`, `spec`, and `selection` methods
 - Added support for Daru data frames
 
 ## 0.1.0 (2020-09-26)