+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (unreleased)
 
 - Added `repeat`, `resolve`, `spec`, and `selection` methods
+- Added `hconcat` and `vconcat` methods
 - Added support for Daru data frames
 
 ## 0.1.0 (2020-09-26)