+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.1 (unreleased)
 
-- Added `repeat`, `resolve`, and `selection` methods
+- Added `repeat`, `resolve`, `spec` and `selection` methods
 - Added `hconcat` and `vconcat` methods
 - Added support for Daru data frames